requirements.txt
certifi=2015.9.6.2
flask==0.10.1
natsort==4.0.4
pillow==2.7.0
python-slugify==1.1.3
requests==2.4.3
  
